refactor: rename aiInterface.ts→aiTextProcessing.ts to distinguish client-side text processing from backend AI service calls
